requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
streamlit==1.42.0
transformers==4.25.1
datasets==3.2.0
donut-python==1.0.9
protobuf==3.20.*
torch==2.6.0
pillow==11.1.0